Interface for types that return a special value for `str()`
Currently the only two types that provide a custom value for `str()` are: string and bytestring. For all others the value is the same as `repr`, as they are both backed by `.String()`.

This special distinction occurs in the repository multiple times, namely in the implementations of `str()`, `print()`, `str.%`, `str.format()`. What they have in common is that they should in fact be backed by whatever value `str()` returns, per the spec.

As such, formalize a `Stringable` interface, implement it only in the `str` type (the behavior of `bytes` remains special) and use it for this distinction rather than directly checking for the type.

This interface finally opens the ability for custom types to have their value of `str()` be distinct from `repr()` (remains backed by `.String()`).